# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Learning Companion Bot API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /learn/progress:
    get:
      operationId: getLearnProgress
      tags: [learn]
      summary: Get saved progress for all topics
      description: Returns all saved topic progress for the given anonymous user
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User's topic progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LearnProgressList"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: saveLearnProgress
      tags: [learn]
      summary: Save progress for a topic
      description: Upserts the completed levels for a topic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveLearnProgressRequest"
      responses:
        "200":
          description: Progress saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LearnProgressItem"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /learn/progress/streak:
    get:
      operationId: getLearnStreak
      tags: [learn]
      summary: Get the current learning streak for a user
      description: Returns the current consecutive-day streak and the longest ever streak
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
          description: The user's unique identifier
      responses:
        "200":
          description: Streak data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LearnStreak"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /learn/progress/activity:
    get:
      operationId: getLearnActivity
      tags: [learn]
      summary: Get activity dates for the heatmap
      description: Returns all dates on which the user had at least one lesson completion, within a given number of past weeks
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
          description: The user's unique identifier
        - name: weeks
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 52
            default: 26
          description: How many weeks of history to return (default 26)
      responses:
        "200":
          description: Activity dates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LearnActivityDates"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    SaveLearnProgressRequest:
      type: object
      properties:
        userId:
          type: string
        language:
          type: string
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced]
        topic:
          type: string
        completedLevels:
          type: array
          items:
            type: integer
      required:
        - userId
        - language
        - difficulty
        - topic
        - completedLevels
    LearnProgressItem:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        language:
          type: string
        difficulty:
          type: string
        topic:
          type: string
        completedLevels:
          type: array
          items:
            type: integer
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - language
        - difficulty
        - topic
        - completedLevels
        - updatedAt
    LearnStreak:
      type: object
      properties:
        currentStreak:
          type: integer
          description: Consecutive days with at least one lesson completed, ending today or yesterday
        longestStreak:
          type: integer
          description: The all-time longest streak in days
        todayActive:
          type: boolean
          description: Whether the user has completed a lesson today
        lastActivityDate:
          type: string
          nullable: true
          description: ISO date string of the most recent activity (YYYY-MM-DD)
      required:
        - currentStreak
        - longestStreak
        - todayActive
        - lastActivityDate
    LearnActivityDates:
      type: object
      properties:
        dates:
          type: array
          items:
            type: string
          description: Array of YYYY-MM-DD strings on which the user had activity
        startDate:
          type: string
          description: The earliest date included in the range (YYYY-MM-DD)
        endDate:
          type: string
          description: The latest date included in the range (YYYY-MM-DD)
      required:
        - dates
        - startDate
        - endDate
    LearnProgressList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/LearnProgressItem"
      required:
        - items
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message