# lib/api-zod/src/generated/types/learnActivityDates.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Learning Companion Bot API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LearnActivityDates {
  /** Array of YYYY-MM-DD strings on which the user had activity */
  dates: string[];
  /** The earliest date included in the range (YYYY-MM-DD) */
  startDate: string;
  /** The latest date included in the range (YYYY-MM-DD) */
  endDate: string;
}