# lib/api-zod/src/generated/types/saveLearnProgressRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Learning Companion Bot API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SaveLearnProgressRequestDifficulty } from "./saveLearnProgressRequestDifficulty";

export interface SaveLearnProgressRequest {
  userId: string;
  language: string;
  difficulty: SaveLearnProgressRequestDifficulty;
  topic: string;
  completedLevels: number[];
}